% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fi_score_plot.R
\name{Fi_score_plot}
\alias{Fi_score_plot}
\title{Fi_score_plot}
\usage{
Fi_score_plot(pdb_df)
}
\arguments{
\item{pdb_df}{Requires a PDB data frame generated by PDB_prepare}
}
\value{
Bar plot
}
\description{
Function to plot Fi-score values per amino acid using a bar plot; NOTE: some PDB files have breakages in their amino acid sequence, that is some residues might be missing and the gaps will be reflected in the plot
}
\examples{
path_to_processed_PDB<- system.file("extdata", "pdb_df.tabular", package="Fiscore")
# basic usage of Fi_score_plot
pdb_df<-read.table(path_to_processed_PDB)
Fi_score_plot(pdb_df)
}
