% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fiscore_secondary.R
\name{Fiscore_secondary}
\alias{Fiscore_secondary}
\title{Fiscore_secondary}
\usage{
Fiscore_secondary(pdb_df)
}
\arguments{
\item{pdb_df}{Requires a PDB data frame generated by PDB_prepare}
}
\value{
bar plot
}
\description{
Function plots a bar plot with  a secondary structure element visualisation based on PDB file data; NOTE: NA refers to unidentified region, e.g., a  likely disordered or unstructured region
}
