% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_psi_bar_plot.R
\name{phi_psi_bar_plot}
\alias{phi_psi_bar_plot}
\title{phi_psi_bar_plot}
\usage{
phi_psi_bar_plot(pdb_df)
}
\arguments{
\item{pdb_df}{Requires a PDB data frame generated by PDB_prepare}
}
\value{
Bar plot
}
\description{
Function to plot PDB file dihedral angle distribution per amino acid using a bar plot. NOTE: some PDB files have breakages in their amino acid sequences; that is, some residues might be missing and it will be reflected in the plot via empty spaces
}
\examples{
path_to_processed_PDB<- system.file("extdata", "pdb_df.tabular", package="Fiscore")
# basic usage of phi_psi_bar_plot
pdb_df<-read.table(path_to_processed_PDB)
phi_psi_bar_plot(pdb_df)
}
