% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_psi_interactive.R
\name{phi_psi_interactive}
\alias{phi_psi_interactive}
\title{phi_psi_interactive}
\usage{
phi_psi_interactive(pdb_df)
}
\arguments{
\item{pdb_df}{Requires a PDB data frame generated by PDB_prepare}
}
\value{
Interactive plot
}
\description{
Function plots a scatter plot with  a secondary structure element visualisation based on the PDB file data; NOTE: NA refers to unidentified region, e.g., a disordered region
}
\examples{
path_to_processed_PDB<- system.file("extdata", "pdb_df.tabular", package="Fiscore")
# basic usage of phi_psi_interactive
pdb_df<-read.table(path_to_processed_PDB)
phi_psi_interactive(pdb_df)
}
