% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_psi_plot.R
\name{phi_psi_plot}
\alias{phi_psi_plot}
\title{phi_psi_plot}
\usage{
phi_psi_plot(pdb_df)
}
\arguments{
\item{pdb_df}{Requires a PDB data frame generated by PDB_prepare}
}
\value{
2D phi/psi angle distribution plot with residues binned
}
\description{
Function to plot a PDB file phi/psi angle density plot
}
\examples{
path_to_processed_PDB<- system.file("extdata", "pdb_df.tabular", package="Fiscore")
# basic usage of phi_psi_plot
pdb_df<-read.table(path_to_processed_PDB)
phi_psi_plot(pdb_df)
}
