% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDB_process.R
\name{PDB_process}
\alias{PDB_process}
\title{PDB_process}
\usage{
PDB_process(file_name, path = "split_PDB")
}
\arguments{
\item{file_name}{PDB file name to load, e.g. '6KZ5.pdb'}

\item{path}{location where to transfer split PDB files, default will create a new directory in your working directory}
}
\value{
generates split chain PDB files in the default or selected directory and then returns the names of the files
}
\description{
Function to preprocess and inspect a PDB file
}
\examples{
path_to_PDB_file<- system.file("extdata", "3nf5.pdb", package="Fiscore")
# basic usage of PDB_process calls the selected path to load a large file
}
